refactor: demos screen logic
