+++ app/navigators/AppNavigator.tsx
@@ -11,13 +11,6 @@ import { AuthNavigator } from "./AuthNavigator"
 import { useStores } from "app/models"
 // import { HudNavigator } from "./HudNavigator"
 
-type DemoScreenList = {
-  ActivityIndicatorDemo: undefined
-  FrameDemo: undefined
-  ButtonDemo: undefined
-  TextInputDemo: undefined
-}
-
 export type AppStackParamList = {
   Auth: undefined
   Hud: undefined
@@ -42,7 +35,8 @@ export type AppStackParamList = {
   ChannelManager: undefined
   RelayManager: undefined
   NotificationSetting: undefined
-} & DemoScreenList
+  Demos: undefined
+}
 
 const exitRoutes = Config.exitRoutes
 
@@ -78,13 +72,7 @@ const AppStack = observer(function AppStack() {
           <Stack.Screen name="ChannelManager" component={Screens.ChannelManagerScreen} />
           <Stack.Screen name="RelayManager" component={Screens.RelayManagerScreen} />
           <Stack.Screen name="NotificationSetting" component={Screens.NotificationSettingScreen} />
-          <Stack.Screen
-            name="ActivityIndicatorDemo"
-            component={Screens.ActivityIndicatorScreenDemo}
-          />
-          <Stack.Screen name="FrameDemo" component={Screens.FrameScreenDemo} />
-          <Stack.Screen name="ButtonDemo" component={Screens.ButtonDemoScreen} />
-          <Stack.Screen name="TextInputDemo" component={Screens.TextInputDemoScreen} />
+          <Stack.Screen name="Demos" component={Screens.DemosScreen} />
         </>
       ) : (
         // <Stack.Screen name="Hud" component={HudNavigator} />
